Feature ETP-5077: Split mocked search flow specs
Checks: npm, testid, tests, regen, xml, pw-mocked, pw-integration
Hooks-Verified: v2 8ee3cd90 522a30074c9a e7e27e9ff429
